Github issue responder/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
PyGithub==2.5.0
anthropic==0.42.0
jinja2==3.1.5
python-multipart==0.0.19
